Rename meta package to metadata to limit breaking changes
