+++ src/main/java/net/minestom/server/item/metadata/PlayerHeadMeta.java
@@ -1,4 +1,4 @@
-package net.minestom.server.item.meta;
+package net.minestom.server.item.metadata;
 
 import net.minestom.server.entity.PlayerSkin;
 import net.minestom.server.item.ItemMeta;
